[processor/remoteobserver] rename to remotetapprocessor
