+++ .github/dependabot.yml
@@ -588,7 +588,7 @@ updates:
       interval: "weekly"
       day: "wednesday"
   - package-ecosystem: "gomod"
-    directory: "/processor/remoteobserverprocessor"
+    directory: "/processor/remotetapprocessor"
     schedule:
       interval: "weekly"
       day: "wednesday"
