+++ cmd/otelcontribcol/processors_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/attraction"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/coreinternal/testdata"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/attributesprocessor"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/remoteobserverprocessor"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/remotetapprocessor"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/resourceprocessor"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/spanprocessor"
 )
@@ -141,7 +141,7 @@ func TestDefaultProcessors(t *testing.T) {
 		{
 			processor: "remoteobserver",
 			getConfigFn: func() component.Config {
-				cfg := procFactories["remoteobserver"].CreateDefaultConfig().(*remoteobserverprocessor.Config)
+				cfg := procFactories["remotetap"].CreateDefaultConfig().(*remotetapprocessor.Config)
 				cfg.Endpoint = "localhost:0"
 				return cfg
 			},
