+++ processor/remotetapprocessor/config_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package remoteobserverprocessor
+package remotetapprocessor
 
 import (
 	"testing"
