+++ processor/remotetapprocessor/factory_test.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package remoteobserverprocessor
+package remotetapprocessor
 
 import (
 	"testing"
@@ -11,7 +11,7 @@ import (
 
 func TestNewFactory(t *testing.T) {
 	factory := NewFactory()
-	assert.EqualValues(t, "remoteobserver", factory.Type())
+	assert.EqualValues(t, "remotetap", factory.Type())
 	config := factory.CreateDefaultConfig()
 	assert.NotNil(t, config)
 }
